backend/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
flask
flask-cors